Update lastfm.py
Thoughts on a small tweak to the last.fm message system? This would send the 'so and so stole the crown' as an embed description rather than a seperate message. What are thoughts on also changing the things like `await ctx.send("Nobody on this server has connected their last.fm account yet!")` as an embed as well rather than a regular message?